+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Volume button goes red when volume is muted
  - Allowing user to route source to 2 destinations without reselecting the source
  - User can switch presentation source without ending presentation and starting again.
+ - VTC Camera close button closes the camera page instead of ending conference
 
 ## [3.2.0] - 2018-05-09
 ### Changed