+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.0.0] - 2018-05-24
+### Added
+ - ATC implementation
+
 ### Changed
  - Default web conferencing path to processor address
  - Volume button goes red when volume is muted